feat: default layout small code refactor
